pkg/cmd: add member add command
